fix(ivy): host bindings should work on nodes with providers (#26771)
PR Close #26771
