+++ packages/core/test/bundling/animation_world/bundle.golden_symbols.json
@@ -171,9 +171,6 @@
     "name": "RENDER_PARENT"
   },
   {
-    "name": "ROOT_EXPANDO_INSTRUCTIONS"
-  },
-  {
     "name": "RecordViewTuple"
   },
   {
