+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -165,9 +165,6 @@
     "name": "RENDER_PARENT"
   },
   {
-    "name": "ROOT_EXPANDO_INSTRUCTIONS"
-  },
-  {
     "name": "RecordViewTuple"
   },
   {
