+++ packages/core/test/bundling/todo_r2/bundle.golden_symbols.json
@@ -699,9 +699,6 @@
     "name": "ROOT_CONTEXT"
   },
   {
-    "name": "ROOT_EXPANDO_INSTRUCTIONS"
-  },
-  {
     "name": "RecordViewTuple"
   },
   {
